Add extra info on Julia support to the community page (#645)
Some extra info on how to get support for Julia was added to the Julia repo as part of
https://github.com/JuliaLang/julia/pull/24052, but we decided that it would be more appropriate to put it
here. I've therefore filled out the Community page slightly with content from that PR. The page now has
more description of what you might use each channel for. It also highlights that newcomers are welcome on
Discourse.