Create JDMJ2017
This paper is of interest to people studying Brownian dynamics and statistical mechanics. We used Julia to do the numerical work and have included several notebooks in the supplemental material explaining our numerical methods. The notebooks rely heavily on DifferentialEquations.jl and JuMP.jl. Although we use Julia throughout the notebooks, we consider this work to be an application of Julia, as opposed to an extension of Julia.